requirements.txt
streamlit
python-dotenv
openai
streamlit-webrtc
av
numpy
SpeechRecognition
pyttsx3
pyaudio
streamlit-mic-recorder
streamlit-audiorecorder
sounddevice
soundfile

